Changing how adding an Event/Workout functions
Basic functionality finished need to implement database queries and then connect EventDetails to the WorkoutDetails => Finally connect to database.
Also UI updates needed
