+++ client/src/pages/AddEvent/AddEventDetails.scss
@@ -1,0 +1,80 @@
+.eventDetails {
+  form {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 100%;
+    flex-direction: column;
+    margin-top: 20px;
+    .list {
+      font-weight: 600;
+      color: black;
+      height: 60%;
+      width: 100%;
+      font-size: 30px;
+      list-style: none;
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      align-items: center;
+
+      div {
+        margin-bottom: 10px;
+        display: flex;
+        align-items: center;
+        flex-direction: column;
+        width: 100%;
+
+        input {
+          height: 25px;
+          position: relative;
+          left: 5px;
+          font-size: 20px;
+        }
+        select {
+          font-size: 20px;
+          width: 25%;
+          text-align: center;
+        }
+        textarea {
+          resize: none;
+          width: 50%;
+          height: 75px;
+          position: relative;
+          left: 5px;
+          font-size: 20px;
+        }
+      }
+      .date {
+        height: 200px;
+        display: flex;
+        flex-direction: row;
+        #datetime-local {
+          width: 80%;
+        }
+      }
+    }
+    .buttonWrapper {
+      display: flex;
+      justify-content: center;
+      width: 100%;
+      button {
+        width: 120px;
+        height: 40px;
+        font-weight: 500;
+        font-size: large;
+        border-radius: 8px;
+        background-color: white;
+        transition: all 0.25s ease;
+        margin: 0px 5px 0px 5px;
+        &:hover {
+          background-color: rgb(219, 219, 219);
+        }
+      }
+    }
+  }
+  &.notActive {
+    display: none;
+  }
+}
