+++ client/src/pages/AddEvent/AddWorkoutDetails.scss
@@ -1,0 +1,52 @@
+.workoutDetails {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  height: 80%;
+  justify-content: space-around;
+  .columnTitles {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    width: 85%;
+    font-size: 30px;
+    font-weight: 600;
+    margin-bottom: 30px;
+    span {
+      width: 30%;
+
+      text-align: center;
+      border-bottom: rgb(87, 87, 87) 3px solid;
+    }
+  }
+  .list {
+    width: 85%;
+    height: 70%;
+    overflow-y: scroll;
+    &::-webkit-scrollbar {
+      display: none;
+    }
+  }
+  .inputs {
+    align-items: center;
+
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    margin-top: 30px;
+    form {
+      width: 85%;
+      display: flex;
+      justify-content: center;
+      input {
+        text-align: center;
+        margin: 5px;
+      }
+    }
+  }
+  &.notActive {
+    display: none;
+  }
+}
